Client/attconfig/requirements.txt
wheel
click
pyfingerprint
mysql.connector
